(BSR)[PRO] refacto: re-organise folder and code
